moved components to components folder. added tabindex. added ids and for to labels and inputs. added required fields
