+++ code/src/components/Footer.js
@@ -3,7 +3,9 @@ import React from 'react';
 const Footer = () => {
     
     return  (
-        <footer className="footer"></footer>
+        <footer className="footer">
+            <p>this is a footer</p>
+        </footer>
     )
 }
