+++ code/src/components/FormContainer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Form from 'Form';
+import Form from 'components/Form';
 
 const FormContainer = () => {
