+++ code/src/components/FormHeader.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 const FormHeader = () => {
     return (
-            <h2 className="form-header">
+            <h2 className="form-header" tabIndex="0">
                 Form
             </h2>
     )
