+++ code/src/components/Header.js
@@ -1,9 +1,10 @@
 import React from 'react';
 
 const Header = () => {
+
     return (
         <header className="header">
-            <h1 className="title">Party</h1>
+            <h1 className="title" tabIndex="0">Party</h1>
         </header>
     )
 }
