+++ code/src/components/Introduction.js
@@ -1,9 +1,10 @@
 import React from 'react';
 
 const Introduction = () => {
+
     return (
         <section className="introduction-container">
-            <p className="introduction">This is an invitation to my birthday party. 
+            <p className="introduction" tabIndex="0">This is an invitation to my birthday party. 
                 Please fill in the following form and click submit.
             </p>
         </section>
