+++ code/src/components/Submit.js
@@ -3,12 +3,13 @@ import React from 'react';
 const Submit = () => {
     
     return (
-        <input 
+        <input
+            id="submit" 
             className="submit-button"
             type="submit"
             value="submit"
             />
-    )
-}
+        )
+    }
 
 export default Submit;
